Move the spec to its new location.
Move README.md to spec/GraphQLOverHTTP.md
About to replace the README.md with a shorter one, but keeping
the history on the spec, per good recommendation from Ivan.

The new spec is in a directory so that it can optionally be
split into separate sections at a later time.