IncrementalDelivery: clarify exactly where CRLF is
According to RFC 1341, there is a CRLF *before* each `--BOUNDARY`, and *before* each payload chunk. That implies there should be a CRLF at the very beginning of the body before the first `--BOUNDARY`. (It also implies that if (as shown in the example) there is no blank link between each JSON chunk and the boundary after it, then the part should not be regarded as containing a trailing newline.)